:truck: Move `_insert_callback` & `callback` into its own function
no other lines were modified during the move
